Tested and works with ORCA. Need to test with a Raymarine control panel or remote.
Adds support to Nav mode (is track in Raymarine)  so we added a ficticious mode in tPyPilotMode of nav. It is sent as compass but allows interptretation of 129184. Also disregards changes in mode in PyPilot except going to disengage!!! (if not the feedback of changing mode to compass stops navigation in Orca.

All navigation is really done by Orca which sends PGN 129184 with heading from current point to next waypoint.

Added correction to variation and deviation to be in degrees in local (NMEA2000 sends them n radians)

Now setCommandheadingTrue takes into account variation and sends the heading - variation as heading to PyPilot.